refactor(data-quality/frame-level): rename review-app → audit-app
Renames the Makefile targets, dep group, Python module, scripts, and
tests from review_app/review-app to audit_app/audit-app. Domain
concepts (reviewer, reviewed status, review.json) are kept since they
describe the act of reviewing, not the app's identity.
