+++ experiments/data-quality/frame-level/Makefile
@@ -1,4 +1,4 @@
-.PHONY: install lint format test help notebook review-app review-export
+.PHONY: install lint format test help notebook audit-app audit-export
 
 help: ## Show this help
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-15s\033[0m %s\n", $$1, $$2}'
@@ -19,11 +19,11 @@ test: ## Run tests with pytest
 notebook: ## Launch JupyterLab for notebooks
 	uv run jupyter lab notebooks/
 
-review-app: ## Launch the bbox-editing review app on http://localhost:8000
-	uv run --group review-app python scripts/run_review_app.py
+audit-app: ## Launch the bbox-editing audit app on http://localhost:8000
+	uv run --group audit-app python scripts/run_audit_app.py
 
-review-export: ## Build YOLO-format patches under data/10_export/ from review.json
-	uv run --group review-app python scripts/export_review_app.py
+audit-export: ## Build YOLO-format patches under data/10_export/ from review.json
+	uv run --group audit-app python scripts/export_audit_app.py
 	@echo ""
 	@echo "Next steps to share the export:"
 	@echo "  uv run dvc add data/10_export && uv run dvc push"
