+++ experiments/data-quality/frame-level/scripts/export_audit_app.py
@@ -12,7 +12,7 @@ from pathlib import Path
 
 import yaml
 
-from data_quality_frame_level.review_app.export_runner import export_one
+from data_quality_frame_level.audit_app.export_runner import export_one
 
 logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
 log = logging.getLogger(__name__)
