+++ experiments/data-quality/frame-level/src/data_quality_frame_level/audit_app/export.py
@@ -8,7 +8,7 @@ The export directory contains four siblings:
   provenance.json       # audit-side context for reproducibility
 
 Each writer here is a pure function over its inputs; the CLI
-(``scripts/export_review_app.py``) gathers the git/DVC/params context
+(``scripts/export_audit_app.py``) gathers the git/DVC/params context
 and feeds it in.
 """
 
@@ -18,9 +18,9 @@ from dataclasses import dataclass
 from datetime import UTC, datetime
 from pathlib import Path
 
+from data_quality_frame_level.audit_app.matching import iou
+from data_quality_frame_level.audit_app.persistence import ReviewState
 from data_quality_frame_level.dataset import BBox
-from data_quality_frame_level.review_app.matching import iou
-from data_quality_frame_level.review_app.persistence import ReviewState
 
 UNCHANGED_IOU = 0.95
 SPURIOUS_MATCH_IOU = 0.95
