+++ experiments/data-quality/frame-level/src/data_quality_frame_level/audit_app/matching.py
@@ -7,8 +7,8 @@ becomes FN.
 
 from dataclasses import dataclass
 
+from data_quality_frame_level.audit_app.types import Prediction
 from data_quality_frame_level.dataset import BBox
-from data_quality_frame_level.review_app.types import Prediction
 
 
 @dataclass(frozen=True)
