+++ experiments/data-quality/frame-level/src/data_quality_frame_level/audit_app/static/app.js
