+++ experiments/data-quality/frame-level/tests/test_audit_app_export.py
@@ -1,8 +1,7 @@
 import json
 from pathlib import Path
 
-from data_quality_frame_level.dataset import BBox
-from data_quality_frame_level.review_app.export import (
+from data_quality_frame_level.audit_app.export import (
     DiffCounts,
     ProvenanceInput,
     compute_diff,
@@ -11,10 +10,11 @@ from data_quality_frame_level.review_app.export import (
     write_pending,
     write_provenance,
 )
-from data_quality_frame_level.review_app.persistence import (
+from data_quality_frame_level.audit_app.persistence import (
     ReviewState,
     SampleReview,
 )
+from data_quality_frame_level.dataset import BBox
 
 
 def _bb(cx, cy, w=0.1, h=0.1):
